Fix incorrect integer type for Matrix row assignment
In practice this worked for x64 however it might cause invalid memory
access on systems with different argument alignment.

+++ source/blender/python/mathutils/mathutils_Matrix.cc
@@ -2432,7 +2432,7 @@ static PyObject *Matrix_item_col(MatrixObject *self, Py_ssize_t col)
 }
 
 /** Sequence accessor (set): `object[i] = x`. */
-static int Matrix_ass_item_row(MatrixObject *self, int row, PyObject *value)
+static int Matrix_ass_item_row(MatrixObject *self, Py_ssize_t row, PyObject *value)
 {
   int col;
   float vec[MATRIX_MAX_DIM];
